python wrapper fix 5
Former-commit-id: 5228ac708eb6028ad660c45dc693f032c8f9a5af [formerly 4aae46fa567adc4e506a5ff514d4d6256eb70ac9] [formerly 2bf3724dc513b116d9f6bcd42b22f0d391115599 [formerly 705452472f6456891a597e514a364dd15f648097]]
Former-commit-id: 8c66a30d5718a9b643ce63567c1dc844dbdd2027 [formerly bddba71d7a880cc4093d95b25a2c13d4209b38cd]
Former-commit-id: b960dc365b30025e8cee23fc91023c02ebe4c949

+++ python/pyfeat.pyx
@@ -88,8 +88,8 @@ cdef class PyFeat:
         cdef np.ndarray[int, ndim=1, mode="fortran"] arr_z_id
         check_X_y(X,y,ensure_2d=True,ensure_min_samples=1)
         X = X.transpose()
-        arr_x = np.asfortranarray(X, dtype=np.float)
-        arr_y = np.asfortranarray(y, dtype=np.float)
+        arr_x = np.asfortranarray(X, dtype=np.float32)
+        arr_y = np.asfortranarray(y, dtype=np.float32)
         arr_z_id = np.asfortranarray(zids, dtype=ctypes.c_int)
 
         self.ft.fit_with_z(&arr_x[0,0],X.shape[0],X.shape[1],&arr_y[0],len(arr_y),
